feat: css수정, 최신순, 인기순 버튼 동작 수정
