+++ src/app/talkroom/page.tsx
@@ -1,11 +1,9 @@
 "use client";
 
-import Arrow from "@/assets/img/arrow.svg";
 import MakeTalkRoom from "@/assets/img/make-talk-room.svg";
 import RecentMakeTalkRoom from "@/assets/img/recent-make-talk-room.svg";
 import { useGetTalkRoomPopular } from "@/hook/reactQuery/main/useGetTalkRoomPopular";
 import { useInput } from "@/hook/useInput";
-import * as DropdownMenu from "@radix-ui/react-dropdown-menu";
 import { usePathname } from "next/navigation";
 import { useState } from "react";
 import { Button } from "../components/Button/Button";
@@ -19,21 +17,15 @@ const page = () => {
   const pathName = usePathname();
 
   const { data: talkRoomPopular } = useGetTalkRoomPopular();
-  const [isStatus, setIsStatus] = useState<boolean>(false);
+  const [status, setStatus] = useState<boolean>(false);
   const { value, handleChange, reset } = useInput("");
-  const [isDate, setIsDate] = useState<string>("날짜별");
 
-  const dateType: string[] = ["~한달 전", "7일전", "하루 전"];
-
-  const changeIsDate = (date: string) => {
-    setIsDate(date);
-  };
   const changeIsStatus = () => {
-    setIsStatus(!isStatus);
+    setStatus(!status);
   };
 
   return (
-    <div>
+    <div className="flex flex-col items-center">
       <div className="w-[1255px]">
         <ThemeMain.MainTheme>
           <div className="flex mt-[78px] mb-[23px]">
@@ -45,51 +37,7 @@ const page = () => {
         </ThemeMain.MainTheme>
         <div className="flex mb-[37px] grow">
           <div className="flex grow">
-            <StatusButton isStatus={isStatus} onClick={changeIsStatus} />
-            <div className="flex items-center ml-[26px]">
-              <DropdownMenu.Root>
-                <DropdownMenu.Trigger asChild>
-                  <button
-                    className="IconButton flex align-center"
-                    aria-label="Customise options"
-                  >
-                    <div className="flex items-center font-Pretendard font-medium text-[#74747B] text-[19px]">
-                      <div className="mr-2">
-                        <Arrow />
-                      </div>
-                      {isDate}
-                    </div>
-                  </button>
-                </DropdownMenu.Trigger>
-
-                <DropdownMenu.Portal>
-                  <DropdownMenu.Content
-                    className="bg-white w-[358px] h-[234px] border border-[#BBB] rounded-[9px] absolute left-[-35px]"
-                    sideOffset={5}
-                  >
-                    <div className="flex text-[21px] font-bold relative border-b-[1px] border-[#D5D5D5] justify-center items-center h-[52px]">
-                      정렬 기준
-                      <div className="absolute right-[5%] cursor-pointer">
-                        X
-                      </div>
-                    </div>
-                    <div className="font-Pretendard font-medium px-[5%] text-[#767676] text-[19px]">
-                      {dateType.map((date, index) => (
-                        <DropdownMenu.Item
-                          key={index}
-                          onClick={() => changeIsDate(date)}
-                          className={`w-full h-[60px] flex items-center px-[5%] hover:font-semibold hover:text-black cursor-pointer ${
-                            date === "하루 전" ? "" : "border-b-[1px]"
-                          }`}
-                        >
-                          {date}
-                        </DropdownMenu.Item>
-                      ))}
-                    </div>
-                  </DropdownMenu.Content>
-                </DropdownMenu.Portal>
-              </DropdownMenu.Root>
-            </div>
+            <StatusButton status={status} onClick={() => changeIsStatus()} />
           </div>
           <div className="flex h-[40px]">
             <div className="w-[567px] mr-[11px]">
